The best solution of Issue #50. Remove the hack.
All of the errors describe in Issue #50 are caused by the garbage
collector.

When the Python interpreter exits it garbage collects those objects that
it can. This means that any corresponding C++ instances and C structures
owned by Python are destroyed. Unfortunately this happens in an
unpredictable order and so can cause memory faults within the wrapped
library.

We used os._exit() as a hack to skip the garbage collection, but it is
not the correct: other cleanup procedure won't run too.

Disable the feature, let us to get rid of it.