Specify the proper size for faces.
To display the faces, FaceWidget creates a WImageLabel and
calls setImageFile(). WImageLabel creates a QMovie and set it
as the image.

But, before the QMovie starts, none of the frame is selected.
So no one knows about the size of the image. So we can't put
them inside a layout.

In the original workaround, we starts it then stop it in order
to get the size information.

But, it is not necessary to do it. We know the size of faces
already. This commit adds two optional parameters to setImageFile(),
we tell the proper size of the faces to WImageLabel, and the overrided
sizeHint() of WImageLabel returns our size.