WImageLabel: free QMovie after stop the animate.
If we don't free QMovie and the users view enough faces by switch tabs
on FaceWindow, we'll soon reach the max open files of the OS limition and
use too much memory.

So, we free QMovie in order to close the file after the animate stop (when
users switch from a tab to another tab in FaceWindow, all faces animates on
the original tab will stop), so we'll never reach the max open files limition.